// Exercises/Part 9/patientor-frontend/src/types.ts
export interface Diagnosis {
	code: string
	name: string
	latin?: string
}

export enum Gender {
	Male = 'male',
	Female = 'female',
	Other = 'other',
}

export interface Patient {
	id: string
	name: string
	occupation: string
	gender: Gender
	ssn?: string
	dateOfBirth?: string
	entries: Array<Entry>
}

export type PatientFormValues = Omit<Patient, 'id' | 'entries'>

export interface Discharge {
	date: string
	criteria: string
}

export interface sickLeave {
	startDate: string
	endDate: string
}

export interface BaseEntry {
	id: string
	description: string
	date: string
	specialist: string
	diagnosisCodes?: Array<Diagnosis['code']>
}

export interface HealthCheckEntry extends BaseEntry {
	type: 'HealthCheck'
	healthCheckRating: HealthCheckRating
}

export interface OccupationalHealthcareEntry extends BaseEntry {
	type: 'OccupationalHealthcare'
	sickLeave: sickLeave
	employerName: string
}

export interface HospitalEntry extends BaseEntry {
	type: 'Hospital'
	discharge: Discharge
}

export enum HealthCheckRating {
	'Healthy' = 0,
	'LowRisk' = 1,
	'HighRisk' = 2,
	'CriticalRisk' = 3,
}

// Define special omit for unions
type UnionOmit<T, K extends string | number | symbol> = T extends unknown
	? Omit<T, K>
	: never
// Define Entry without the 'id' property
export type EntryWithoutId = UnionOmit<Entry, 'id'>

export type Entry =
	| HospitalEntry
	| OccupationalHealthcareEntry
	| HealthCheckEntry